Extracted library from package
