+++ packages/simulator_2d/CMakeLists.txt
@@ -11,13 +11,51 @@ find_package(tf2_geometry_msgs REQUIRED)
 find_package(tf2_msgs REQUIRED)
 find_package(rosgraph_msgs REQUIRED)
 
+if(NOT CMAKE_CXX_STANDARD)
+    set(CMAKE_CXX_STANDARD 20)
+endif()
+
 add_compile_options(-Wall -Wextra -Wpedantic -Werror)
+
+add_library(simulator_engine SHARED src/simulator_engine.cpp)
+
+ament_target_dependencies(
+    simulator_engine
+    rclcpp
+    model
+    geom)
+
+target_include_directories(
+    simulator_engine PUBLIC
+    "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
+    "$<INSTALL_INTERFACE:include/${PROJECT_NAME}>"
+)
+
+ament_export_targets(simulator_engine_targets HAS_LIBRARY_TARGET)
+
+ament_export_dependencies(
+    rclcpp
+    model
+    geom)
+
+install(
+  TARGETS simulator_engine
+  EXPORT simulator_engine_targets
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
+)
+
+install(
+  DIRECTORY include/
+  DESTINATION include/${PROJECT_NAME}
+)
 
 add_executable(
     ${PROJECT_NAME}_node
-    src/main.cpp
+    src/simulator_engine.cpp
     src/simulator_node.cpp
-    src/simulator_engine.cpp)
+    src/main.cpp)
 
 ament_target_dependencies(
     ${PROJECT_NAME}_node
@@ -30,25 +68,19 @@ ament_target_dependencies(
     tf2_msgs
     rosgraph_msgs)
 
-if(BUILD_TESTING)
-    find_package(ament_cmake_gtest REQUIRED)
-    ament_add_gtest(${PROJECT_NAME}_test test/main.cpp WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
-    target_link_libraries(${PROJECT_NAME}_test ${PROJECT_NAME})
-endif()
-
 install(
     TARGETS ${PROJECT_NAME}_node
     DESTINATION lib/${PROJECT_NAME})
 
 install(
-    DIRECTORY launch
-    DESTINATION share/${PROJECT_NAME})
-
-target_include_directories(
-    ${PROJECT_NAME}_node PUBLIC
-    "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
-    "$<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/include>"
-    "$<INSTALL_INTERFACE:include/${PROJECT_NAME}>"
+  DIRECTORY launch
+  DESTINATION share/${PROJECT_NAME}
 )
+
+if(BUILD_TESTING)
+    find_package(ament_cmake_gtest REQUIRED)
+    ament_add_gtest(${PROJECT_NAME}_test test/main.cpp WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
+    target_link_libraries(${PROJECT_NAME}_test simulator_engine)
+endif()
 
 ament_package()
