+++ packages/simulator_2d/src/main.cpp
@@ -1,6 +1,6 @@
-#include "simulator_2d/simulator_node.h"
+#include "simulator_node.h"
 
-#include "rclcpp/rclcpp.hpp"
+#include <rclcpp/rclcpp.hpp>
 
 int main(int argc, char* argv[]) {
     rclcpp::init(argc, argv);
